pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ros2vlx"
version = "0.1.0"
description = "ROS2 VLX - A natural language interface for ROS2."
authors = [
    { name = "Manuel Schweiger", email = "schweigermanuel@icloud.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "langchain-nvidia-ai-endpoints",
    "langchain-community>=0.3.1",
    "langchain-core>=0.3.9",
    "langchain-openai>=0.2.2",
    "langchain-ollama>=0.2.0",
    "python-dotenv>=0.21.0",
    "pyyaml>=6.0",
    "setuptools>=42",
    "wheel>=0.37.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "black>=22.0",
    "flake8>=5.0"
]

[project.urls]
Homepage = "https://github.com/mschweig"
Repository = "https://github.com/mschweig/ros2vlx"
Issues = "https://github.com/mschweig/ros2vlx/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["ros2vlx*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v --cov=ros2vlx --cov-report=html"
testpaths = ["tests"]